backend/pyproject.toml
[project]
name = "ghost-search"
version = "0.1.0"
description = "AI-powered privacy-first media search backend for Ghost Downloader"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "langchain-core>=0.3.0",
    "langchain-groq>=0.2.0",
]

[project.optional-dependencies]
vector = ["chromadb>=0.5.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
